requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.8
django-cors-headers==3.0.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.1
django-import-export==1.2.0
django-rest-elasticsearch==0.4.1
django-reversion==3.0.3
django-stubs==0.12.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
elasticsearch==6.3.1
elasticsearch-dsl==6.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
mypy==0.701
mypy-extensions==0.4.1
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.2
Pillow==4.2.1
PyJWT==1.7.1
pytest-runner==4.4
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
typed-ast==1.3.5
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.3
xadmin2==2.0.1
xlrd==1.2.0
XlsxWriter==1.1.7
xlwt==1.3.0
